Time: 618 ms (5.09%), Space: 33.7 MB (93.43%) - LeetHub

+++ 0480-sliding-window-median/0480-sliding-window-median.cpp
@@ -3,26 +3,23 @@ public:
     vector<double> medianSlidingWindow(vector<int>& nums, int k) {
         vector<double> res;
         vector<long long> ans;
-        
-        for(int i = 0; i < k;i++){
-            ans.insert(lower_bound(ans.begin(),ans.end(),nums[i]),nums[i]);
+        for(int i = 0 ; i < k; i++){
+            ans.insert(lower_bound(ans.begin(), ans.end(), nums[i]), nums[i]);
         }
-        
-        if(k%2==0){
-            res.push_back((double) (ans[k/2] + ans[k/2 - 1]) / 2.0);
-        } else{
-            res.push_back((double)ans[k/2]);
+        if(k % 2==0){
+            res.push_back( (double) (ans[k/2] + ans[k/2-1])/ 2.0);
+        } else {
+            res.push_back(ans[k/2]);
         }
-
-        for(int i = k; i < nums.size();i++){
-            ans.erase(lower_bound(ans.begin(), ans.end(), nums[i-k]));
-            ans.insert(lower_bound(ans.begin(),ans.end(),nums[i]),nums[i]);
-            if(k%2==0){
-                res.push_back((double) (ans[k/2] + ans[k/2 - 1]) / 2.0);
-            } else{
+        for(int i = k ; i < nums.size();i++){
+            ans.erase(lower_bound(ans.begin(), ans.end(),nums[i-k]));
+            ans.insert(lower_bound(ans.begin(), ans.end(), nums[i]), nums[i]);
+            if(k % 2==0){
+                res.push_back( (double) (ans[k/2] + ans[k/2-1])/ 2.0);
+            } else {
                 res.push_back(ans[k/2]);
             }            
         }
-        return res;
+       return res;
     }
 };
